requirement.txt
numpy
opencv-python
Pillow
ImageHash
ultralytics
paddleocr
paddlepaddle>=2.6.0
torch>=2.0.0
torchvision>=0.15.0
git+https://github.com/KaiyangZhou/deep-person-reid.git
